reigistry user is more generic (#15)
We can set the REGISTRY_USER as a group variable to the value samsung_cnct. This will automatically apply the variable to all projects in the group. Third party users can either insert their org name in the file, or settheir own group vars.
No extra setup is needed, although we should make a note in solas that this is being done.